ci(toolkit): use latest GitHub Actions Toolkit API to safely set environment variables
A moderate security vulnerability has been identified in the GitHub Actions runner that can allow environment variable and path injection in workflows that log untrusted data to STDOUT. This can result in environment variables being introduced or modified without the intention of the workflow author. Action authors who are using the toolkit should update the @actions/core package to v1.2.6 or greater to get the updated addPath and exportVariable functions.

For more information see https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/.

closes #28