feat(toolkit/sashbox): prevent the splitter from overlapping overlays
In the sashbox component, we form a new stacking context to make
nested stacking contexts local, preventing nested stacking contexts
from leaking the sashbox element boundary.

We use `0` as its z-index value. A z-index of `0` is like setting no z-index
(auto) at all, except that it still forms a stacking context. Since it has the
order `0`, it will not overlap subsequent positioned elements of the same
parent stacking context.

Our rationale for setting a z-index on the splitter:
The splitter is a positioned element; thus, it implicitly forms a stacking
context that overlaps the subsequent sash (as long as the sash is statically
positioned). But sash content may also contain positioned elements or
even have elements with a z-index set. To prevent them from overlapping
the splitter, we form a stacking context on the sash div by setting its
z-index to `0` and set the splitter’s z-index to `1`.