feat(toolkit/observable): allow observing an element's position on the screen
The Observable returned uses a relatively simple approach to detecting element position changes. It assumes the browser to support the native {ResizeObserver} API or logs a warning otherwise.

There is, unfortunately, no native browser API to detect position changes of an element in a performant and reliable way. Our approach to detecting position changes is based on the premise that in most cases, an element's position changes when the size of a parent element changes (or one of the parent's direct child elements), or if they are scrolled, or when elements are added to or removed from the DOM. This approach may cover most use cases, but not all, e.g., when moving an element without causing the size of its container element to change.

We are aware that this approach can be quite expensive, mainly because potentially a large number of elements need to be monitored for resizing/scrolling. Therefore, use this Observable only when absolutely necessary, and simple size changes are not sufficient.