feat(toolkit/viewport): allow a viewport to grow with its content
We changed the viewport to grow with its content, allowing for implementing use cases where the viewport grows with its content up to a maximum height and then starts overflowing its content.

BREAKING CHANGE: Changed the `sci-viewport` element to grow with its content. The viewport no longer positions its content absolutely, but in the document element flow instead, allowing the viewport to adapt its size to its content's width and height. This change may be breaking. Please refer to our documentation on how to layout a viewport: https://github.com/SchweizerischeBundesbahnen/scion-toolkit/blob/master/docs/site/tools/viewport.md.

As part of this change, we also had to change the `sci-scrollable` directive's behavior. It no longer positions its host absolutely, but in document element flow instead. You may use this directive for adding the SCION scrollbar to other viewport implementations, such as the `<cdk-virtual-scroll-viewport>` component of Angular CDK. To migrate, add the host to a CSS grid container to fill the remaining space vertically and horizontally.