ci: run ngcc before building the testing app in `before-push` script
When building (or starting) the testing app without first explicitly compiling the toolkit libraries to the Ivy format, the testing app's compilation fails with the error NG6002 because `ngcc` does not wait for all secondary entry points to finish compilation.

Note that the SCION Workbench project, which also consists of multiple library projects, does not have that issue as not exporting to secondary entry points.

closes #49