feat(toolkit/splitter): provide splitter Angular component to control the size of elements next to it
The `<sci-splitter>` Angular component is a visual element that allows the user to control the size of elements next to it.
It has a handle that the user can move depending on the orientation of the splitter.

In this commit, we changed the `sciSplitter` directive, which was an internal of `SciSashboxComponent`, to a public Angular component.

For more information, refer to https://github.com/SchweizerischeBundesbahnen/scion-toolkit/blob/master/docs/site/tools/splitter.md