fix(toolkit/viewport): remove DOM elements instantly after computing the native scrollbar track size
To compute the track size of native scrollbars, the `SciNativeScrollbarTrackSizeProvider` service creates a `<div>` viewport element and adds it to the DOM. After the computation, it removes it again. However, since the computation runs outside the Angular zone, `Renderer2` does not remove the element instantly, but on the next change detection cycle, resulting in many `<div>` elements being added to the document body.

To avoid having to synchronize with the Angular zone, we changed the implementation to add/remove DOM elements directly using the DOM API instead of `Renderer2` API.