refactor(ɵcomponents): remove page objects of internal components
The page objects of the internal components are moved to the end-to-end
tests of the projects which use them. Otherwise, these projects would have to
configure ESM for running Playwright end-to-end tests. As ESM is not yet well
supported by most IDEs, this would lead to a poor developer experience.

closes #78