fix(toolkit/bean-manager): do not destroy "value beans"
Previously, the bean manager checked for the presence of a "preDestroy" method for beans that were registered as value beans and, if present, called it upon destruction. Value beans, however, should not be destroyed at all. For example, registering a Window of a different origin would result in a cross-origin access error when checking for the presence of such a method.