fix(components/viewport): consider elements as scrolled into view when there is no viewport overflow
The calculation of whether an element is scrolled into view may be wrong by a few pixels if the viewport contains elements with decimal sizes. This can happen because `offsetLeft` and `offsetTop` operate on an integer (not a decimal), losing precision that can accumulate.

To avoid incorrect calculations, the viewport assumes that its elements are scrolled into view when there is no viewport overflow.