fix(components/viewport): do not throw error when calling `computeOffset` for an element not contained in the viewport
The viewport now returns `null` instead of throwing an error when calling `computeOffset` for an element that is not contained in the viewport or if its effective `display` style resolves to `none`.

BREAKING CHANGE: more tolerant handling of elements not contained in the viewport introduced a breaking change.

The following methods of `SciViewportComponent` are now more tolerant when invoked for elements that are not contained in the viewport or have `display` style set to `none`.

- `SciViewportComponent#computeOffset` returns `null` instead of throwing an error.
- `SciViewportComponent#isElementInView` returns `false` instead of throwing an error.
- `SciViewportComponent#scrollIntoView` does nothing instead of throwing an error.