feat(toolkit/operators): provide replacement for `subscribeInside` and `observeInside` operators
The new operators `subscribeIn` and `observeIn` replace the `subscribeInside` and `observeInside` operators.

The problem with `subscribeInside` is that it also wraps downstream operators, which can be confusing and lead to unexpected behavior, e.g., if the function continues execution asynchronously. To set up a context for both, subscription and downstream operators, use `subscribeIn` in combination with the `observeIn` operator.